+++ Tests/RxTest.swift
@@ -88,6 +88,7 @@ extension RxTest {
 
     func setUpActions(){
         _ = Hooks.defaultErrorHandler // lazy load resource so resource count matches
+        _ = Hooks.subscriptionCallstackGenerationHandler // lazy load resource so resource count matches
         #if TRACE_RESOURCES
             self.startResourceCount = Resources.total
             //registerMallocHooks()